+++ include/mapnik/json/extract_bounding_box_grammar_impl.hpp
@@ -148,13 +148,10 @@ extract_bounding_box_grammar<Iterator, Boxes, ErrorHandler>::extract_bounding_bo
     json.value = json.object | json.array | json.string_ | json.number
         ;
 
-    json.pairs = json.key_value % lit(',')
+    json.key_value = json.string_ >> lit(':') >> json.value
         ;
 
-    json.key_value = (json.string_ >> lit(':') >> json.value)
-        ;
-
-    json.object = lit('{') >> *json.pairs >> lit('}')
+    json.object = lit('{') >>  json.key_value % lit(',') >> lit('}')
         ;
 
     json.array = lit('[')