+++ include/mapnik/json/topojson_grammar.hpp
@@ -42,6 +42,138 @@ namespace qi = boost::spirit::qi;
 namespace fusion = boost::fusion;
 using space_type = mapnik::json::space_type;
 
+struct create_point
+{
+    using result_type = mapnik::topojson::point;
+    template <typename T0, typename T1>
+    result_type operator()(T0 & coord, T1 & props) const
+    {
+        mapnik::topojson::point pt;
+        if (coord.template is<mapnik::topojson::coordinate>())
+        {
+            auto const& coord_ = coord.template get<mapnik::topojson::coordinate>();
+            pt.coord = coord_;
+            pt.props = props;
+        }
+        return pt;
+    }
+};
+
+struct create_multi_point
+{
+    using result_type = mapnik::topojson::multi_point;
+    template <typename T0, typename T1>
+    result_type operator()(T0 & coords, T1 & props) const
+    {
+        mapnik::topojson::multi_point mpt;
+        if (coords.template is<std::vector<mapnik::topojson::coordinate>>())
+        {
+            auto const& points = coords.template get<std::vector<mapnik::topojson::coordinate>>();
+            mpt. points = points;
+            mpt.props = props;
+        }
+        return mpt;
+    }
+};
+
+struct create_line_string
+{
+    using result_type = mapnik::topojson::linestring;
+    template <typename T0, typename T1>
+    result_type operator()(T0 & arcs, T1 & props) const
+    {
+        mapnik::topojson::linestring line;
+        if (arcs.template is<std::vector<index_type>>())
+        {
+            auto const& arcs_ = arcs.template get<std::vector<index_type>>();
+            line.rings = arcs_;
+            line.props = props;
+        }
+        return line;
+    }
+};
+
+struct create_multi_line_string
+{
+    using result_type = mapnik::topojson::multi_linestring;
+    template <typename T0, typename T1>
+    result_type operator()(T0 & arcs, T1 & props) const
+    {
+        mapnik::topojson::multi_linestring mline;
+        if (arcs.template is<std::vector<std::vector<index_type>>>())
+        {
+            auto const& arcs_ = arcs.template get<std::vector<std::vector<index_type>>>();
+            mline.lines = arcs_;
+            mline.props = props;
+        }
+        return mline;
+    }
+};
+
+struct create_polygon
+{
+    using result_type = mapnik::topojson::polygon;
+    template <typename T0, typename T1>
+    result_type operator()(T0 & arcs, T1 & props) const
+    {
+        mapnik::topojson::polygon poly;
+        if (arcs.template is<std::vector<std::vector<index_type>>>())
+        {
+            auto const& arcs_ = arcs.template get<std::vector<std::vector<index_type>>>();
+            poly.rings = arcs_;
+            poly.props = props;
+        }
+        return poly;
+    }
+};
+
+struct create_multi_polygon
+{
+    using result_type = mapnik::topojson::multi_polygon;
+    template <typename T0, typename T1>
+    result_type operator()(T0 & arcs, T1 & props) const
+    {
+        mapnik::topojson::multi_polygon mpoly;
+        if (arcs.template is<std::vector<std::vector<std::vector<index_type>>>>())
+        {
+            auto const& arcs_ = arcs.template get<std::vector<std::vector<std::vector<index_type>>>>();
+            mpoly.polygons = arcs_;
+            mpoly.props = props;
+        }
+        return mpoly;
+    }
+};
+
+
+struct create_geometry_impl
+{
+    using result_type = mapnik::topojson::geometry;
+    template <typename T0, typename T1, typename T2, typename T3>
+    result_type operator()(T0 geom_type, T1 & coord, T2 & arcs, T3 & props) const
+    {
+        switch (geom_type)
+        {
+        case 1: //Point
+            return create_point()(coord, props);
+        case 2: //LineString
+            return create_line_string()(arcs, props);
+        case 3: //Polygon
+            return create_polygon()(arcs, props);
+        case 4: //MultiPoint
+            return create_multi_point()(coord, props);
+        case 5: //MultiLineString
+            return create_multi_line_string()(arcs, props);
+        case 6: //MultiPolygon
+            return create_multi_polygon()(arcs, props);
+        default:
+            break;
+        }
+        return mapnik::topojson::geometry(); //empty
+    }
+};
+
+using coordinates_type = util::variant<coordinate,std::vector<coordinate>>;
+using arcs_type = util::variant<std::vector<index_type>, std::vector<std::vector<index_type>>, std::vector<std::vector<std::vector<index_type>>>>;
 template <typename Iterator, typename ErrorHandler = json::error_handler<Iterator> >
 struct topojson_grammar : qi::grammar<Iterator, space_type, topology()>
 
@@ -55,24 +187,18 @@ private:
     qi::rule<Iterator, space_type, std::vector<mapnik::topojson::geometry>()> objects;
     qi::rule<Iterator, space_type, std::vector<mapnik::topojson::arc>()> arcs;
     qi::rule<Iterator, space_type, mapnik::topojson::arc()> arc;
-    qi::rule<Iterator, space_type, mapnik::topojson::coordinate()> coordinate;
+    qi::rule<Iterator, space_type, mapnik::topojson::coordinate()> coordinate_;
+    qi::rule<Iterator, space_type, coordinates_type()> coordinates;
     qi::rule<Iterator, space_type, mapnik::topojson::transform()> transform;
     qi::rule<Iterator, space_type, mapnik::topojson::bounding_box()> bbox;
-    qi::rule<Iterator, space_type, mapnik::topojson::geometry() > geometry;
-    qi::rule<Iterator, space_type, mapnik::topojson::point()> point;
-    qi::rule<Iterator, space_type, mapnik::topojson::multi_point()> multi_point;
-    qi::rule<Iterator, space_type, mapnik::topojson::linestring()> linestring;
-    qi::rule<Iterator, space_type, mapnik::topojson::multi_linestring()> multi_linestring;
-    qi::rule<Iterator, space_type, mapnik::topojson::polygon()> polygon;
-    qi::rule<Iterator, space_type, mapnik::topojson::multi_polygon()> multi_polygon;
+    qi::rule<Iterator, qi::locals<int, coordinates_type, arcs_type, properties>, mapnik::topojson::geometry(), space_type> geometry;
     qi::rule<Iterator, space_type, void(std::vector<mapnik::topojson::geometry>&)> geometry_collection;
     qi::rule<Iterator, space_type, std::vector<index_type>()> ring;
+    qi::rule<Iterator, space_type, std::vector<std::vector<index_type>>()> rings;
+    qi::rule<Iterator, space_type, arcs_type()> rings_array;
     // properties
-    qi::rule<Iterator, space_type, mapnik::topojson::properties()> properties;
-    qi::rule<Iterator, space_type, mapnik::topojson::properties()> attributes;
-    qi::rule<Iterator, space_type, mapnik::json::json_value()> attribute_value;
-    // id
-    qi::rule<Iterator,space_type> id;
+    qi::rule<Iterator, space_type, mapnik::topojson::properties()> properties_;
+    qi::symbols<char, int> geometry_type_dispatch;
 };
 
 }}