+++ src/geometry_envelope.cpp
@@ -38,8 +38,6 @@ template MAPNIK_DECL mapnik::box2d<double> envelope(multi_line_string<double> co
 template MAPNIK_DECL mapnik::box2d<double> envelope(multi_polygon<double> const& geom);
 // collection
 template MAPNIK_DECL mapnik::box2d<double> envelope(geometry_collection<double> const& geom);
-//
-template MAPNIK_DECL mapnik::box2d<std::int64_t> envelope(geometry<std::int64_t> const& geom);
 
 } // end ns geometry
 } // end ns mapnik