+++ src/proj_transform.cpp
@@ -200,20 +200,21 @@ bool proj_transform::backward (double * x, double * y , double * z, int point_co
 #ifdef MAPNIK_USE_PROJ4
     if (is_dest_longlat_)
     {
-        int i;
-        for(i=0; i<point_count; i++) {
-            x[i*offset] *= DEG_TO_RAD;
-            y[i*offset] *= DEG_TO_RAD;
+        for (int i = 0; i < point_count; ++i)
+        {
+            x[i * offset] *= DEG_TO_RAD;
+            y[i * offset] *= DEG_TO_RAD;
         }
     }
 
-    if (pj_transform( dest_.proj_, source_.proj_, point_count,
-                      offset, x,y,z) != 0)
+    if (pj_transform(dest_.proj_, source_.proj_, point_count,
+                     offset, x, y, z) != 0)
     {
         return false;
     }
 
-    for(int j=0; j<point_count; j++) {
+    for (int j = 0; j < point_count; ++j)
+    {
         if (x[j] == HUGE_VAL || y[j] == HUGE_VAL)
         {
             return false;
@@ -222,10 +223,10 @@ bool proj_transform::backward (double * x, double * y , double * z, int point_co
 
     if (is_source_longlat_)
     {
-        int i;
-        for(i=0; i<point_count; i++) {
-            x[i*offset] *= RAD_TO_DEG;
-            y[i*offset] *= RAD_TO_DEG;
+        for (int i = 0; i < point_count; ++i)
+        {
+            x[i * offset] *= RAD_TO_DEG;
+            y[i * offset] *= RAD_TO_DEG;
         }
     }
 #endif
@@ -266,7 +267,7 @@ unsigned int proj_transform::backward (std::vector<geometry::point<double>> & ls
     double * x = reinterpret_cast<double*>(ptr);
     double * y = x + 1;
     double * z = nullptr;
-    if(!backward(x, y, z, size, 2))
+    if (!backward(x, y, z, size, 2))
     {
         return size;
     }
@@ -312,31 +313,24 @@ bool proj_transform::backward (box2d<double> & box) const
     if (is_source_equal_dest_)
         return true;
 
-    double llx = box.minx();
-    double ulx = box.minx();
-    double lly = box.miny();
-    double lry = box.miny();
-    double lrx = box.maxx();
-    double urx = box.maxx();
-    double uly = box.maxy();
-    double ury = box.maxy();
-    double z = 0.0;
-    if (!backward(llx,lly,z))
+    double x[4], y[4];
+    x[0] = box.minx(); // llx 0
+    y[0] = box.miny(); // lly 1
+    x[1] = box.maxx(); // lrx 2
+    y[1] = box.miny(); // lry 3
+    x[2] = box.minx(); // ulx 4
+    y[2] = box.maxy(); // uly 5
+    x[3] = box.maxx(); // urx 6
+    y[3] = box.maxy(); // ury 7
+
+    if (!backward(x, y, nullptr, 4, 1))
         return false;
-    if (!backward(lrx,lry,z))
-        return false;
-    if (!backward(ulx,uly,z))
-        return false;
-    if (!backward(urx,ury,z))
-        return false;
-    double minx = std::min(ulx, llx);
-    double miny = std::min(lly, lry);
-    double maxx = std::max(urx, lrx);
-    double maxy = std::max(ury, uly);
-    box.init(minx,
-             miny,
-             maxx,
-             maxy);
+
+    double minx = std::min(x[0], x[2]);
+    double miny = std::min(y[0], y[1]);
+    double maxx = std::max(x[1], x[3]);
+    double maxy = std::max(y[2], y[3]);
+    box.init(minx, miny, maxx, maxy);
     return true;
 }
 
@@ -425,7 +419,6 @@ bool proj_transform::backward(box2d<double>& env, int points) const
     env.re_center(result.center().x, result.center().y);
     env.height(result.height());
     env.width(result.width());
-
     return true;
 }