+++ test/unit/datasource/topojson.cpp
@@ -21,6 +21,7 @@
  *****************************************************************************/
 
 #include "catch.hpp"
+#include "ds_test_util.hpp"
 
 #include <mapnik/util/fs.hpp>
 #include <mapnik/util/file_io.hpp>
@@ -50,7 +51,7 @@ bool parse_topology(std::string const& filename, mapnik::topojson::topology & to
 
 }
 
-TEST_CASE("topology")
+TEST_CASE("topojson")
 {
     SECTION("geometry parsing")
     {
@@ -72,4 +73,42 @@ TEST_CASE("topology")
             }
         }
     }
+
+    SECTION("TopoJSON properties are properly expressed")
+    {
+        std::string filename("./test/data/topojson/escaped.topojson");
+        mapnik::context_ptr ctx = std::make_shared<mapnik::context_type>();
+        mapnik::transcoder tr("utf8");
+        mapnik::topojson::topology topo;
+        REQUIRE(parse_topology(filename, topo));
+        mapnik::value_integer feature_id = 0;
+        for (auto const& geom : topo.geometries)
+        {
+            mapnik::box2d<double> bbox = mapnik::util::apply_visitor(mapnik::topojson::bounding_box_visitor(topo), geom);
+            CHECK(bbox.valid());
+            mapnik::topojson::feature_generator<mapnik::context_ptr> visitor(ctx, tr, topo, feature_id);
+            mapnik::feature_ptr feature = mapnik::util::apply_visitor(visitor, geom);
+            CHECK(feature);
+            CHECK(feature->envelope() == bbox);
+            std::initializer_list<attr> attrs = {
+                attr{"name", tr.transcode("Test")},
+                attr{"NOM_FR", tr.transcode("Québec")},
+                attr{"boolean", mapnik::value_bool("true")},
+                attr{"description", tr.transcode("Test: \u005C")},
+                attr{"double", mapnik::value_double(1.1)},
+                attr{"int", mapnik::value_integer(1)},
+                attr{"object", tr.transcode("{name:\"waka\",spaces:\"value with spaces\",int:1,double:1.1,boolean:false"
+                                                            ",NOM_FR:\"Québec\",array:[\"string\",\"value with spaces\",3,1.1,null,true"
+                                                            ",\"Québec\"],another_object:{name:\"nested object\"}}")},
+                attr{"spaces", tr.transcode("this has spaces")},
+                attr{"array", tr.transcode("[\"string\",\"value with spaces\",3,1.1,null,true,"
+                                                           "\"Québec\",{name:\"object within an array\"},"
+                                                           "[\"array\",\"within\",\"an\",\"array\"]]")},
+                attr{"empty_array", tr.transcode("[]")},
+                attr{"empty_object", tr.transcode("{}")},
+            };
+            REQUIRE_ATTRIBUTES(feature, attrs);
+        }
+    }
+
 }